Desk balance is full integers

+++ js/collectibles-dapp.js
@@ -13,7 +13,7 @@ function TncDapp() {
 
         $(document).ready(async function(){
             let balance = await web3.eth.getBalance ( tncLib.account );
-            balance = parseFloat(_this.formatNumberString(balance, 18))
+            balance = parseInt(_this.formatNumberString(balance, 18))
             
             //Parsing float to remove trailing zeroes.      
             $('#wallet-balance').text(balance)
